Fix issue with Eigen memory alignment
libolfsysm.a was being compiled with -march=native, which can enable
instructions requiring specific memory alignment and thereby change
Eigen's behavior regarding memory allocation. However the R binding
olfsysm.so was not compiled with -march=native, so the Eigen code
'native' to olfsysm.so was unaware of the altered memory-related
behavior used in the Eigen code from libolfsysm.a, causing crashes
whenever memory allocated in libolfsysm.a was freed by olfsysm.so.